remove old linkTo
