--- src/libs/Navigation/newLinkTo/types.ts
@@ -1,11 +1,0 @@
-type ActionPayloadParams = {
-    screen?: string;
-    params?: unknown;
-    path?: string;
-};
-
-type ActionPayload = {
-    params?: ActionPayloadParams;
-};
-
-export type {ActionPayload, ActionPayloadParams};
